Replace pid with taskid in task_g_pre_launch_priv signature
The parameters passed to this function were the job and the pid, but the job
was not really used in task plugin. Replace the pid with the taskid and make
plugins to obtain the pid from the job and taskid.

This is a preparation for the task plugin to be able to work at a task level
possibly creating and constraining per task cgroup directories.

There's no change in functionality with this commit.

Bug 10375