Make cgroup structure for jobacctgather more generic
This plugin is different from the others because it uses the step_x/task_y
structure in the cgroup directories for the cpuacct and memory controllers.

With this in mind the initial design was to create two functions of cgroup
accounting and delegate the creation/removal of steps and task directories
for these two controllers to the cgroup plugin. This is not flexible
because we are integrating the decision of which controllers to monitor into
the cgroup plugin instead of the jobacctgather plugin.

Given the new features required in bug 10375 this needs to change now.

This patch does just remove these specific functions:

cgroup_g_accounting_init
cgroup_g_accounting_fini
cgroup_g_accounting_task_addto_accounting

And adds:

cgroup_g_task_addto

The logic of the previous functions is now split into the jobacctgather
cgroup plugin and into these two new functions, and the destruction of the
created task_x directories is delegated to the cgroup_p_step_destroy.

The advantatges will be that now any plugin could make use of task_addto and
for a specific controller to add a step_x/task_y directory, and constrain or
get statistics at task level. The cleanup is done by the last plugin which
calls cgroup_p_step_destroy for the given controller.

Note also that extern step pids are not added to a task, so we needed to
create a new function for this.

These changes are a previous preparation for bug 10375 where we need to
constrain devices to specific tasks.

Bug 10375