Fix setting cpuset.mems and cpuset.cpus on certain OSes
In certain OSes there's a situation where cpuset.cpus and cpuset.mems are
set inheriting parent's cgroup parameters, but in others they are just left
blank. Treat this situation to correctly set the values in all cases which
will also avoid an EBUSY error when trying to write to an incorrectly set
cgroup.

Bug 12157