PE-38620: fix switch plugin CXI setup
PE-38620 - Slurm: fix Slingshot switch plugin for nodes with
  non-working NICs

Allow empty slots in the cxi_devs[] array, if opening a device
fails.  Check for empty slots when creating service profiles
for an application, and do a device-name search when mapping
a profile to a slot in cxi_devs[].
Lastly, change some error messages to include slot numbers
and device names.

Tested on sawmill with nodes that have failed NICs.